+++ src/ml_board/frontend/newDashboard/src/components/dashboard/table/Table.scss
