+++ src/ml_board/frontend/newDashboard/src/components/graphs/Graphs.css
