+++ src/ml_board/frontend/newDashboard/src/components/settings/Settings.tsx
@@ -1,6 +1,8 @@
 function Settings() {
   return (
-    'settings'
+    <div>
+      Settings
+    </div>
   )
 }
