+++ src/ml_board/frontend/newDashboard/src/components/throughputs/Throughput.tsx
@@ -1,6 +1,8 @@
 function Throughput() {
   return (
-    'throughput'
+    <div>
+      Throughput
+    </div>
   )
 }
