--- src/ml_board/frontend/newDashboard/src/redux/actions/ExperimentActions.tsx
@@ -1,8 +1,0 @@
-export const EVAL_RESULT = 'EVAL_RESULT';
-
-export function saveEvalResultData(evalResultData: any){
-    return {
-        type: EVAL_RESULT,
-        evalResultData: evalResultData
-    }
-}
