--- src/ml_board/frontend/newDashboard/src/ui/alerts/Alert.tsx
@@ -1,7 +1,0 @@
-function Alert() {
-    return (
-      'Alert'
-    )
-  }
-  
-  export default Alert;
