--- src/ml_board/frontend/newDashboard/src/ui/alerts/Alerts.tsx
@@ -1,7 +1,0 @@
-function Alerts() {
-    return (
-      'Alerts'
-    )
-  }
-  
-  export default Alerts;
