--- src/ml_board/frontend/newDashboard/src/ui/chart/Charts.scss
@@ -1,40 +1,0 @@
-.charts {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, calc(50% - 15px));
-    grid-gap: 5px;
-    height: 100%;
-    width: 100%;
-    padding: 10px;
-    box-sizing: border-box;
-    overflow: auto;
-
-    .chartTitle {
-        width: 100%;
-        place-self: right;
-        align-self: center;
-        transform: translateX(-10px);
-        font-size: 0.65em;
-        color: rgb(175, 180, 184);
-        text-align: end;
-        z-index: 2;
-
-        .titlePart-0 {
-            margin-right: 5px;
-            border: 1px solid grey;
-            padding: 1px 3px;
-            color: white;
-            background: darkgrey;
-            border-radius: 5px;
-            z-index: 2;
-        }
-
-        .titlePart-1 {
-            border: 1px solid rgb(69, 126, 108);
-            padding: 1px 3px;
-            color: rgb(70, 67, 67);
-            background: rgb(153, 179, 161);
-            border-radius: 5px;
-            z-index: 2;
-        }
-    }
-}
