--- src/ml_board/frontend/newDashboard/src/ui/routes/NewCharts.tsx
@@ -1,9 +1,0 @@
-import AppNew from '../../app/AppNew';
-
-function NewCharts() {
-  return (
-    <AppNew/>
-  )
-}
-
-export default NewCharts;
