--- src/ml_board/frontend/newDashboard/src/ui/table/Table.tsx
@@ -1,59 +1,0 @@
-import 'ag-grid-community/styles/ag-grid.css'; // Core grid CSS, always needed
-import 'ag-grid-community/styles/ag-theme-alpine.css'; // Optional theme CSS
-import { AgGridReact } from "ag-grid-react";
-// styles
-import './Table.scss';
-
-
-interface columnDefinition {
-  field: string;
-  sortable?: boolean;
-  filter?: boolean
-}
-
-
-function Table({ colNames, rows }: { colNames: string[], rows: any[] }) {
-
-  // change the array of strings to array of colum definitions
-  const colDefs: columnDefinition[] = colNames.map(
-    (colName: string) => ({
-      field: colName,
-      sortable: true,
-      filter: true
-    })
-  );
-
-  return (
-    <div className="ag-theme-alpine" id="ag-grid-container">
-      <AgGridReact
-        defaultColDef={{ resizable: true }}
-        // {/* provide column definitions */}
-        columnDefs={colDefs}
-        // {/* specify auto group column definition */}
-        // autoGroupColumnDef={this.autoGroupColumnDef}
-        // {/* row data provided via props from the file store */}
-        rowData={rows}
-        // // enable tree data
-        // treeData={true}
-        // // {/* return tree hierarchy from supplied data */}
-        // getDataPath={data => data.filePath}
-        // // {/* expand tree by default */}
-        // groupDefaultExpanded={-1}
-        // // {/* fit grid columns */}
-        onGridReady={params => params.api.sizeColumnsToFit()}
-        // // {/* provide context menu callback */}
-        // getContextMenuItems={this.getContextMenuItems}
-        // // {/* provide row drag end callback */}
-        // onRowDragEnd={this.onRowDragEnd}
-        // // {/* return id required for tree data and immutable data */}
-        getRowId={(params: any) => params.data.job_id}
-        // // {/* specify our FileCellRenderer component */}
-        // components={this.components}
-        animateRows={true} // Optional - set to 'true' to have rows animate when sorted
-      >
-      </AgGridReact>
-    </div>
-  )
-}
-
-export default Table;
